chore(chatbar): cursor class
