+++ components/views/chat/chatbar/Chatbar.html
@@ -8,7 +8,6 @@
     :class="{'is-error': charlimit, 'inactive': !recipient, 'top-corners-sharp': showFilePreview || ui.replyChatbarContent.id || commandPreview}"
     ref="wrap"
   >
-
     <div class="control-wrap file-upload-wrap">
       <FilesUpload
         :editable="recipient ? true : false"
